refactor either and imports
